feat: Don't make this field full width by default
This change is up for discussion. The issue is, that
SupportsFullWidth does not allow to unset the
fullWidth flag but only set it.
By passing the vue prop properly, we gain control
over whether we _want_ the field to be full width or
not, however in most cases the answer probably is yes
and this would lead to additionall ->fullWidth()
invocations having to be added to all Medialibrary
field uses among a lot of projects.